Fix issues

+++ datagridview/Forms/MainForm.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Threading.Tasks;
 using System.Windows.Forms;
 using datagridview.Contracts;
 using datagridview.Forms;
@@ -21,8 +22,6 @@ namespace datagridview
 
             bindingSource = new BindingSource();
             dataGridView1.DataSource = bindingSource;
-
-            UpdateToolStrip();
         }
 
         private async void btnEdit_Click(object sender, EventArgs e)
@@ -33,11 +32,14 @@ namespace datagridview
             }
             var selectedTour = (Contracts.Models.Tour)dataGridView1.Rows[dataGridView1.SelectedRows[0].Index].DataBoundItem;
             var form = new EditTourForm(selectedTour);
-            form.ShowDialog();
+            if (form.ShowDialog() != DialogResult.OK)
+            {
+                return;
+            }
 
             await tourManager.EditTourAsync(form.CurrentTour);
             bindingSource.ResetBindings(false);
-            UpdateToolStrip();
+            await UpdateToolStrip();
         }
 
         private async void btnAdd_Click(object sender, EventArgs e)
@@ -50,7 +52,7 @@ namespace datagridview
 
             await tourManager.AddTourAsync(form.CurrentTour);
             bindingSource.ResetBindings(false);
-            UpdateToolStrip();
+            await UpdateToolStrip();
         }
 
         private async void btnDelete_Click(object sender, EventArgs e)
@@ -60,53 +62,53 @@ namespace datagridview
                 return;
             }
 
+            DialogResult result = MessageBox.Show(
+                "Вы действительно удалить строку?",
+                "Подтверждение удаления",
+                MessageBoxButtons.YesNo,
+                MessageBoxIcon.Question
+            );
+
+            if (result != DialogResult.Yes)
+            {
+                return;
+            }
+
             var selectedTour = (Contracts.Models.Tour)dataGridView1.CurrentRow.DataBoundItem;
 
             await tourManager.DeleteTourAsync(selectedTour.Id);
             bindingSource.ResetBindings(false);
-            UpdateToolStrip();
+            await UpdateToolStrip();
         }
 
-        private void UpdateToolStrip()
+        private async Task UpdateToolStrip()
         {
-            var toursCount = dataGridView1.RowCount;
-            var sum = 0.0;
-            var toursCountWithAdd = 0;
-            var additionSum = 0.0;
+            var stats = await tourManager.GetStatsAsync();
 
-            for (var i = 0; i < dataGridView1.RowCount; i++)
-            {
-                var tour = (Contracts.Models.Tour)dataGridView1.Rows[i].DataBoundItem;
-                sum += tour.TotalCost;
-
-                if (tour.AdditionalFees != 0)
-                {
-                    ++toursCountWithAdd;
-                    additionSum += tour.AdditionalFees;
-                }
-            }
-
-            strip1.Text = $"Количество туров {toursCount};";
-            strip2.Text = $"Общая сумма {sum}; ";
-            strip3.Text = $"Количество туров с доплатами {toursCountWithAdd}; ";
-            strip4.Text = $"Общая сумма доплат {additionSum}; ";
+            strip1.Text = $"Количество туров {stats.ToursCount};";
+            strip2.Text = $"Общая сумма {stats.TotalSum}; ";
+            strip3.Text = $"Количество туров с доплатами {stats.ToursCountWithAdd}; ";
+            strip4.Text = $"Общая сумма доплат {stats.AdditionSum}; ";
         }
 
         private async void MainForm_Load(object sender, EventArgs e)
         {
             bindingSource.DataSource = await tourManager.GetAllToursAsync();
             dataGridView1.Columns[nameof(Contracts.Models.Tour.Id)].Visible = false;
+            dataGridView1.Columns.Add("TotalCost", "Общая стоимость");
+
+            await UpdateToolStrip();
         }
 
-        private void dataGridView1_CellFormatting(object sender, DataGridViewCellFormattingEventArgs e)
+        private async void dataGridView1_CellFormatting(object sender, DataGridViewCellFormattingEventArgs e)
         {
             if (dataGridView1.Columns[e.ColumnIndex].Name == "TotalCost")
             {
                 var data = (Contracts.Models.Tour)dataGridView1.Rows[e.RowIndex].DataBoundItem;
-                data.TotalCost = data.PricePerTour * data.PeopleCount + data.AdditionalFees;
-                e.Value = data.TotalCost;
+                dataGridView1[e.ColumnIndex, e.RowIndex].Value = data.PricePerTour * data.PeopleCount + data.AdditionalFees;
+                e.Value = dataGridView1[e.ColumnIndex, e.RowIndex].Value;
             }
-            UpdateToolStrip();
+            await UpdateToolStrip();
         }
     }
 }
